requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
langchain
openai
tavily-python
langchain_openai
asyncpg

